refactor(agents): move interrupted-state finalizer out of the vendored stream port
finalize_interrupted_ui_message_state has no upstream counterpart, so it
belongs in ui_message_stream (the Phoenix-extensions module) rather than
the faithful AI SDK port in vercel_ui_message_stream.
